fix(editor) Explicitly prevent vscode actions updating vscode
