+++ editor/src/components/editor/action-types.ts
@@ -26,7 +26,6 @@ import {
   NodeModules,
   Imports,
   ParsedTextFile,
-  HighlightBoundsForUids,
   ImageFile,
 } from '../../core/shared/project-file-types'
 import { CodeResultCache, PropertyControlsInfo } from '../custom-code/code-file'
@@ -47,7 +46,6 @@ import {
   ElementsToRerender,
   ErrorMessages,
   FloatingInsertMenuState,
-  GithubRepo,
   GithubState,
   LeftMenuTab,
   ModalDialog,
@@ -57,15 +55,12 @@ import {
   RightMenuTab,
   StoredEditorState,
   GithubOperation,
-  FileChecksums,
   GithubData,
   UserConfiguration,
   ThemeSetting,
   ColorSwatch,
-  NavigatorEntry,
 } from './store/editor-state'
 import { Notice } from '../common/notice'
-import { UtopiaVSCodeConfig } from 'utopia-vscode-common'
 import type { LoginState } from '../../common/user'
 import { InsertableComponent, StylePropOption } from '../shared/project-components'
 import { LayoutTargetableProp } from '../../core/layout/layout-helpers-new'
@@ -76,6 +71,7 @@ import { CanvasCommand } from '../canvas/commands/commands'
 import { InsertionPath } from './store/insertion-path'
 import { TextProp } from '../text-editor/text-editor'
 import { ElementPathTrees } from '../../core/shared/element-path-tree'
+import { FromVSCodeAction } from './actions/actions-from-vscode'
 export { isLoggedIn, loggedInUser, notLoggedIn } from '../../common/user'
 export type { LoginState, UserDetails } from '../../common/user'
 
@@ -679,13 +675,6 @@ export interface UpdateFromWorker {
   updates: Array<WorkerParsedUpdate | WorkerCodeAndParsedUpdate>
 }
 
-export interface UpdateFromCodeEditor {
-  action: 'UPDATE_FROM_CODE_EDITOR'
-  filePath: string
-  savedContent: string
-  unsavedContent: string | null
-}
-
 export interface ClearParseOrPrintInFlight {
   action: 'CLEAR_PARSE_OR_PRINT_IN_FLIGHT'
 }
@@ -726,12 +715,6 @@ export interface SetCodeEditorBuildErrors {
 export interface SetCodeEditorLintErrors {
   action: 'SET_CODE_EDITOR_LINT_ERRORS'
   lintErrors: ErrorMessages
-}
-
-export interface SendLinterRequestMessage {
-  action: 'SEND_LINTER_REQUEST_MESSAGE'
-  filePath: string
-  content: string
 }
 
 export interface SaveDOMReport {
@@ -900,22 +883,6 @@ export interface UpdateText {
   textProp: TextProp
 }
 
-export interface MarkVSCodeBridgeReady {
-  action: 'MARK_VSCODE_BRIDGE_READY'
-  ready: boolean
-}
-
-export interface SelectFromFileAndPosition {
-  action: 'SELECT_FROM_FILE_AND_POSITION'
-  filePath: string
-  line: number
-  column: number
-}
-
-export interface SendCodeEditorInitialisation {
-  action: 'SEND_CODE_EDITOR_INITIALISATION'
-}
-
 export interface SetFocusedElement {
   action: 'SET_FOCUSED_ELEMENT'
   focusedElementPath: ElementPath | null
@@ -935,11 +902,6 @@ export interface SetScrollAnimation {
 export interface SetFollowSelectionEnabled {
   action: 'SET_FOLLOW_SELECTION_ENABLED'
   value: boolean
-}
-
-export interface UpdateConfigFromVSCode {
-  action: 'UPDATE_CONFIG_FROM_VSCODE'
-  config: UtopiaVSCodeConfig
 }
 
 export interface SetLoginState {
@@ -1030,15 +992,6 @@ export interface SetResizeOptionsTargetOptions {
   action: 'SET_RESIZE_OPTIONS_TARGET_OPTIONS'
   propertyTargetOptions: Array<LayoutTargetableProp>
   index: number | null
-}
-
-export interface HideVSCodeLoadingScreen {
-  action: 'HIDE_VSCODE_LOADING_SCREEN'
-}
-
-export interface SetIndexedDBFailed {
-  action: 'SET_INDEXED_DB_FAILED'
-  indexedDBFailed: boolean
 }
 
 export interface ForceParseFile {
@@ -1187,7 +1140,6 @@ export type EditorAction =
   | UpdateGithubData
   | RemoveFileConflict
   | UpdateFromWorker
-  | UpdateFromCodeEditor
   | ClearParseOrPrintInFlight
   | ClearImageFileBlob
   | AddFolder
@@ -1196,7 +1148,6 @@ export type EditorAction =
   | SetMainUIFile
   | SetCodeEditorBuildErrors
   | SetCodeEditorLintErrors
-  | SendLinterRequestMessage
   | SaveDOMReport
   | SetProp
   | SetPropWithElementPath
@@ -1226,14 +1177,10 @@ export type EditorAction =
   | UpdatePropertyControlsInfo
   | AddStoryboardFile
   | UpdateText
-  | MarkVSCodeBridgeReady
-  | SelectFromFileAndPosition
-  | SendCodeEditorInitialisation
   | SetFocusedElement
   | ScrollToElement
   | SetScrollAnimation
   | SetFollowSelectionEnabled
-  | UpdateConfigFromVSCode
   | SetLoginState
   | SetGithubState
   | SetUserConfiguration
@@ -1250,8 +1197,6 @@ export type EditorAction =
   | DecrementResizeOptionsSelectedIndex
   | IncrementResizeOptionsSelectedIndex
   | SetResizeOptionsTargetOptions
-  | HideVSCodeLoadingScreen
-  | SetIndexedDBFailed
   | ForceParseFile
   | RunEscapeHatch
   | SetElementsToRerender
@@ -1266,6 +1211,7 @@ export type EditorAction =
   | SetConditionalOverriddenCondition
   | SwitchConditionalBranches
   | UpdateConditionalExpression
+  | FromVSCodeAction
 
 export type DispatchPriority =
   | 'everyone'
