+++ editor/src/components/editor/store/editor-update.tsx
@@ -225,6 +225,7 @@ export function runSimpleLocalEditorAction(
     case 'ADD_FOLDER':
       return UPDATE_FNS.ADD_FOLDER(action, state)
     case 'DELETE_FILE':
+    case 'DELETE_FILE_FROM_VSCODE':
       return UPDATE_FNS.DELETE_FILE(action, state, derivedState, userState)
     case 'ADD_TEXT_FILE':
       return UPDATE_FNS.ADD_TEXT_FILE(action, state)
