+++ editor/src/components/editor/store/vscode-changes.ts
@@ -32,6 +32,7 @@ import {
   getUnderlyingVSCodeBridgeID,
 } from './editor-state'
 import { shallowEqual } from '../../../core/shared/equality-utils'
+import * as EP from '../../../core/shared/element-path'
 
 export interface WriteProjectFileChange {
   type: 'WRITE_PROJECT_FILE'
@@ -186,10 +187,10 @@ export function shouldIncludeVSCodeDecorations(
     [...newEditorState.highlightedViews, ...newEditorState.selectedViews],
     newEditorState,
   )
-  return (
-    oldEditorState.selectedViews !== newEditorState.selectedViews ||
-    oldEditorState.highlightedViews !== newEditorState.highlightedViews ||
-    !shallowEqual(oldHighlightBounds, newHighlightBounds)
+  return !(
+    EP.arrayOfPathsEqual(oldEditorState.selectedViews, newEditorState.selectedViews) &&
+    EP.arrayOfPathsEqual(oldEditorState.highlightedViews, newEditorState.highlightedViews) &&
+    shallowEqual(oldHighlightBounds, newHighlightBounds)
   )
 }
 
@@ -206,9 +207,10 @@ export function shouldIncludeSelectedElementChanges(
     newEditorState,
   )
   return (
-    (oldEditorState.selectedViews !== newEditorState.selectedViews ||
-      !shallowEqual(oldHighlightBounds, newHighlightBounds)) &&
-    newEditorState.selectedViews.length > 0
+    !(
+      EP.arrayOfPathsEqual(oldEditorState.selectedViews, newEditorState.selectedViews) &&
+      shallowEqual(oldHighlightBounds, newHighlightBounds)
+    ) && newEditorState.selectedViews.length > 0
   )
 }
